// src/api/nexuscore/model/eventOut.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */
import type { CountdownBlock } from './countdownBlock';
import type { CtaBannerBlock } from './ctaBannerBlock';
import type { DividerBlock } from './dividerBlock';
import type { EventOutStatus } from './eventOutStatus';
import type { FaqBlock } from './faqBlock';
import type { HeroBlock } from './heroBlock';
import type { HighlightGridBlock } from './highlightGridBlock';
import type { MediaGalleryBlock } from './mediaGalleryBlock';
import type { NarrativeBlock } from './narrativeBlock';
import type { PrizeTiersBlock } from './prizeTiersBlock';
import type { RulesColumnsBlock } from './rulesColumnsBlock';
import type { StatRowBlock } from './statRowBlock';

export interface EventOut {
  /** The url slug of the event */
  slug: string;
  /** The title of the event */
  title: string;
  /** A short description of the event */
  description: string;
  /** The URL of the card image */
  image_url: string | null;
  /** The start time of the event */
  start_time: string;
  /** The end time of the event */
  end_time: string;
  /** The status of the event */
  status: EventOutStatus;
  /** The blocks used for the event page display */
  blocks: (HeroBlock | CountdownBlock | StatRowBlock | NarrativeBlock | HighlightGridBlock | PrizeTiersBlock | RulesColumnsBlock | FaqBlock | MediaGalleryBlock | CtaBannerBlock | DividerBlock)[];
}